+++ Changelog.md
@@ -9,17 +9,32 @@
 to specify a density.
     * [Pull request 248](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/248)
 
+1. Added a Sphere class.
+    * [Pull request 255](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/255)
+
+1. Added a `MassMatrix3::SetFromSphere` function that uses a `Material` to
+specify a density.
+    * [Pull request 247](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/247)
+
 1. Added a `MassMatrix3::SetFromBox` function that uses a `Material` to specify
    a density.
     * [Pull request 246](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/246)
 
 1. Deprecated mutator functions in MassMatrix3 that lacked a `Set` prefix.
-    * [Pull request 246](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/246)
+    * [Pull request 262](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/262)
 
 
 ## Ignition Math 4.x
 
 ### Ignition Math 4.x.x
+
+1. Add Graph::EdgeFromVertices function that return an edge, if one exists,
+   between two vertices.
+    * [Pull request 254](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/254)
+
+1. Added multiply assign operator to Matrix4. 
+    * [Pull request 252](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/252)
+
 
 1. Add Plane copy constructor and fix cppcheck on artful
     * [Pull request 230](https://bitbucket.org/ignitionrobotics/ign-math/pull-requests/230)